+++ packages/core/.storybook/mocks/cart-sidebar.js
@@ -1,0 +1,40 @@
+export const cartItem = {
+  id: '39542524',
+  price: 656.5,
+  listPrice: 892.43,
+  seller: {
+    identifier: '1',
+  },
+  quantity: 1,
+  itemOffered: {
+    sku: '39542524',
+    name: 'pink',
+    gtin: '0001078884042',
+    image: [
+      {
+        url: 'https://storeframework.vtexassets.com/arquivos/ids/165082/sed.jpg?v=637752996157770000',
+        alternateName: 'accusamus',
+      },
+      {
+        url: 'https://storeframework.vtexassets.com/arquivos/ids/165079/quis.jpg?v=637752996135400000',
+        alternateName: 'dolorem',
+      },
+      {
+        url: 'https://storeframework.vtexassets.com/arquivos/ids/165080/tempora.jpg?v=637752996141030000',
+        alternateName: 'enim',
+      },
+      {
+        url: 'https://storeframework.vtexassets.com/arquivos/ids/165081/est.jpg?v=637752996144600000',
+        alternateName: 'ea',
+      },
+    ],
+    brand: {
+      name: 'iRobot',
+    },
+    isVariantOf: {
+      productGroupID: '99483893',
+      name: 'Tasty Concrete Computer',
+    },
+    additionalProperty: [],
+  },
+}